+++ src/main/java/gift/dto/request/AuthRequest.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.request;
 
 public class AuthRequest {
     private String email;
